bridge: br_fill_forward_path add check for _UNTAG_HW only
Passing path = NULL, then only check for switchdev egress.

When the port and vlan id stored in ctx was added by switchdev,
then we return with path->num_vlans = 1.

It is used during evaluation of the flow, when examining the skb for
encapsulated vlans. We need to know if the vlan tag found in the skb is
added by the switchdev.