COVERLETTER roam
title=flow offload teardown when layer 2 roaming
prefix=PATCH v2 nf-next
remcc=linux-kernel@vger.kernel.org
remcc=coreteam@netfilter.org

/*** BLURB HERE ***/

In case of a bridge in the forward-fastpath or bridge-fastpath the fdb is
used to create the tuple. In case of roaming at layer 2 level, for example
802.11r, the destination device is changed in the fdb. The destination
device of a direct transmitting tuple is no longer valid and traffic is
send to the wrong destination. Also the hardware offloaded fastpath is not
valid anymore.

This flowentry needs to be torn down asap. Also make sure that the flow
entry is not being used, when marked for teardown.

Changes in v2:
- Unchanged, only tags RFC net-next to PATCH nf-next.